agent/requirements.txt
pytest
paramiko
requests
allure-pytest
python-dotenv